pyproject.toml
[project]
name = "dex-sim"
version = "0.1.0"
description = "A minimalistic deterministic simulation for leverage, margin, and CCP logic"
authors = [{ name = "Pedro Argento" }]
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "arch==8.0.0",
    "statsmodels>=0.13.0",
    "scipy>=1.8",
    "matplotlib>=3.7",
    "seaborn>=0.13.2",
    "numba>=0.62.1",
    "zarr>=2.18.3",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
plot = []

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sim*"]

[tool.ruff]
line-length = 100

[project.scripts]
dex-sim = "dex_sim.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

# (Optional) include non-Python files
[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml"]
